eventstore module is separated to sub project for cohesion.
