+++ eventstore/build.gradle
@@ -1,0 +1,4 @@
+plugins {
+    id 'java-library'
+}
+
