+++ eventstore/src/main/java/net/slipp/ddd/events/DomainEvent.java
@@ -1,9 +1,8 @@
+package net.slipp.ddd.events;
 /**
  * Created by joenggyu0@gmail.com on 4/28/20
  * Github : http://github.com/lenkim
  */
-
-package net.slipp.common.domain.model;
 
 import java.time.LocalDateTime;
