+++ eventstore/src/main/java/net/slipp/ddd/events/DomainEventPublisher.java
@@ -1,7 +1,4 @@
-package net.slipp.common.domain.model;
-
-import net.slipp.utils.Assertions;
-
+package net.slipp.ddd.events;
 import java.util.ArrayList;
 import java.util.List;
