+++ recruit/src/main/java/net/slipp/franchise/domain/model/recruit/Recruit.java
@@ -1,7 +1,7 @@
 package net.slipp.franchise.domain.model.recruit;
 
 import com.google.common.collect.Sets;
-import net.slipp.common.domain.model.DomainEventPublisher;
+import net.slipp.ddd.events.DomainEventPublisher;
 import net.slipp.franchise.domain.model.user.UserId;
 import net.slipp.utils.Assertions;
